refactor: separate route parsing utilities
