+++ src/pages.ts
@@ -5,7 +5,8 @@ import { parse as parseSFC, compileScript } from '@vue/compiler-sfc'
 import { walk } from 'estree-walker'
 import { mkdir, readFile, writeFile } from 'node:fs/promises'
 import MagicString from 'magic-string'
-import { formatMessage, getRoutePath, parseSegment, readFileSync } from './utils'
+import { formatMessage, readFileSync } from './utils'
+import { getRoutePath, parseSegment } from './utils/route-parsing'
 import { localizeRoutes } from './routing'
 import { mergeLayerPages } from './layers'
 import { resolve, parse as parsePath, dirname } from 'pathe'
